% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title lab 8}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{process}

\item{iter}{iteration}

\item{lambda}{lambda}

\item{...}{table variables}
}
\value{
root
}
\description{
Title lab 8
}
\examples{
mycltp(5,10000)
}
