% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Title Binomial Simulation}
\usage{
myf(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{trial of success}

\item{p}{probability of a success}
}
\value{
a sophisticated plot
}
\description{
Title Binomial Simulation
}
\examples{
myf(100,10,0.5)
}
