% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Title Find Max likelihood}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{log lik is a combination bin}

\item{theta}{theta}
}
\value{
max likelihood
}
\description{
Title Find Max likelihood
}
